In the auto item edit window, clicking the change button was giving an invalid cast error.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3235 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Forms/FormAutoItemEdit.cs
@@ -213,11 +213,11 @@ namespace OpenDental{
 				return;
       }
 			if(AutoCodeItemC.HList.ContainsKey(FormP.SelectedCodeNum)
-				&& (int)AutoCodeItemC.HList[FormP.SelectedCodeNum] != AutoCodeItemCur.AutoCodeNum)
+				&& (long)AutoCodeItemC.HList[FormP.SelectedCodeNum] != AutoCodeItemCur.AutoCodeNum)
 			{
 				//This section is a fix for an old bug that did not cause items to get deleted properly
-				if(!AutoCodeC.HList.ContainsKey((int)AutoCodeItemC.HList[FormP.SelectedCodeNum])){
-					AutoCodeItems.Delete((int)AutoCodeItemC.HList[FormP.SelectedCodeNum]);
+				if(!AutoCodeC.HList.ContainsKey((long)AutoCodeItemC.HList[FormP.SelectedCodeNum])){
+					AutoCodeItems.Delete((long)AutoCodeItemC.HList[FormP.SelectedCodeNum]);
 					textADA.Text=ProcedureCodes.GetStringProcCode(FormP.SelectedCodeNum);
 				}
 				else{
